google_poc/requirements.txt
google-cloud-speech
google-cloud-texttospeech
google-generativeai